python-receiver-code/requirements.txt
numpy>=1.21
vispy>=0.16.0
glfw>=2.10.0
pyserial>=3.5
